Delete jsonldcontext.json
I had to remove this to do the merge.   The lines
```
 "ecrro": "http://purl.obolibrary.org/obo/ECRRO_",
        "ecrr": "https://n2t.net/ark:/23942/g2",
```
would need to go in a EC context file.  We can't modify the schema.org context to support these.  We can make an EC context file if we need to, but that might be something to do in the Science on Schema repo and not Gleaner's

Thanks for this!   I really appreciate your contributions.